+++ src/app/services/archive-project.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Observable, Subject } from 'rxjs';
+import { Observable, Subject, Subscription } from 'rxjs';
 import { Project } from '../domain/project';
 import { ArchiveProjectResponse } from '../domain/archiveProjectResponse';
 import { MatSnackBar } from '@angular/material/snack-bar';
@@ -16,7 +16,7 @@ export class ArchiveProjectService {
     return this.http.put<ArchiveProjectResponse>(`/api/project/${project.id}/archived`, null);
   }
 
-  archiveProjects(projects: Project[]): Observable<ArchiveProjectResponse[]> {
+  makeArchiveProjectsRequest(projects: Project[]): Observable<ArchiveProjectResponse[]> {
     const projectIds = projects.map((project) => project.id);
     return this.http.put<ArchiveProjectResponse[]>(`/api/projects/archived`, projectIds);
   }
@@ -25,7 +25,7 @@ export class ArchiveProjectService {
     return this.http.delete<ArchiveProjectResponse>(`/api/project/${project.id}/archived`);
   }
 
-  unarchiveProjects(projects: Project[]): Observable<ArchiveProjectResponse[]> {
+  makeUnarchiveProjectsRequest(projects: Project[]): Observable<ArchiveProjectResponse[]> {
     let params = new HttpParams();
     for (const project of projects) {
       params = params.append('projectIds', project.id);
@@ -65,12 +65,6 @@ export class ArchiveProjectService {
     this.refreshProjects();
   }
 
-  updateProjectsArchivedStatus(projects: Project[], archived: boolean): void {
-    projects.forEach((project) => {
-      this.updateProjectArchivedStatus(project, archived);
-    });
-  }
-
   showArchiveProjectsSuccessMessage(projects: Project[], archived: boolean): void {
     this.snackBar
       .open(
@@ -81,7 +75,10 @@ export class ArchiveProjectService {
       )
       .onAction()
       .subscribe(() => {
-        this.undoArchiveProjects(projects, archived ? 'unarchiveProjects' : 'archiveProjects');
+        this.undoArchiveProjects(
+          projects,
+          archived ? 'makeUnarchiveProjectsRequest' : 'makeArchiveProjectsRequest'
+        );
       });
   }
 
@@ -104,5 +101,73 @@ export class ArchiveProjectService {
     this.snackBar.open(
       archive ? $localize`Error archiving unit.` : $localize`Error restoring unit.`
     );
+  }
+
+  archiveProjects(projects: Project[], archive: boolean): Subscription {
+    return this[archive ? 'makeArchiveProjectsRequest' : 'makeUnarchiveProjectsRequest'](
+      projects
+    ).subscribe({
+      next: (archiveProjectsResponse: ArchiveProjectResponse[]) => {
+        this.updateProjectsArchivedStatus(projects, archiveProjectsResponse);
+        this.openSuccessSnackBar(projects, archiveProjectsResponse, archive);
+      },
+      error: () => {
+        this.showErrorSnackBar(archive);
+      }
+    });
+  }
+
+  private updateProjectsArchivedStatus(
+    projects: Project[],
+    archiveProjectsResponse: ArchiveProjectResponse[]
+  ): void {
+    for (const archiveProjectResponse of archiveProjectsResponse) {
+      const project = projects.find((project: Project) => project.id === archiveProjectResponse.id);
+      project.updateArchivedStatus(archiveProjectResponse.archived);
+    }
+    this.refreshProjects();
+  }
+
+  private openSuccessSnackBar(
+    projects: Project[],
+    archiveProjectsResponse: ArchiveProjectResponse[],
+    archived: boolean
+  ): void {
+    const count = archiveProjectsResponse.filter(
+      (response: ArchiveProjectResponse) => response.archived === archived
+    ).length;
+    this.snackBar
+      .open(
+        archived
+          ? $localize`Successfully archived ${count} unit(s).`
+          : $localize`Successfully restored ${count} unit(s).`,
+        $localize`Undo`
+      )
+      .onAction()
+      .subscribe(() => {
+        this.undoArchiveAction(
+          projects,
+          archived ? 'makeUnarchiveProjectsRequest' : 'makeArchiveProjectsRequest'
+        );
+      });
+  }
+
+  private showErrorSnackBar(archive: boolean): void {
+    this.snackBar.open(
+      archive ? $localize`Error archiving unit(s).` : $localize`Error restoring unit(s).`
+    );
+  }
+
+  private undoArchiveAction(projects: Project[], archiveFunctionName: string): void {
+    this[archiveFunctionName](projects).subscribe({
+      next: (archiveProjectsResponse: ArchiveProjectResponse[]) => {
+        this.updateProjectsArchivedStatus(projects, archiveProjectsResponse);
+        this.refreshProjects();
+        this.snackBar.open($localize`Action undone.`);
+      },
+      error: () => {
+        this.snackBar.open($localize`Error undoing action.`);
+      }
+    });
   }
 }
